+++ src/constants.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_PROTOCOL = 757;
+export const DEFAULT_PORT = 25565;
+export const DEFAULT_TIMEOUT = 3600;
