+++ src/ping-strategy.ts
@@ -1,0 +1,4 @@
+export abstract class PingStrategy {
+  abstract createHandshakePacket(): Buffer;
+  abstract parse(response: Buffer): IPingResponse;
+}
